Fix up 'merge master'

+++ c/column.h
@@ -337,8 +337,8 @@ public:
   void replace_buffer(MemoryRange&&) override;
   const T* elements_r() const;
   T* elements_w();
-  T get_elem(int64_t i) const;
-  void set_elem(int64_t i, T value);
+  T get_elem(size_t i) const;
+  void set_elem(size_t i, T value);
 
   size_t data_nrows() const override;
   void resize_and_fill(size_t nrows) override;
@@ -364,7 +364,7 @@ protected:
 };
 
 
-template <> void FwColumn<PyObject*>::set_elem(int64_t, PyObject*);
+template <> void FwColumn<PyObject*>::set_elem(size_t, PyObject*);
 extern template class FwColumn<int8_t>;
 extern template class FwColumn<int16_t>;
 extern template class FwColumn<int32_t>;
